requirements.txt
aio-pika==9.4.3
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.2.0
black==24.10.0
boto3==1.35.20
botocore==1.35.56
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.0
dnspython==2.7.0
dramatiq==1.17.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
orjson==3.10.11
packaging==24.2
pamqp==3.3.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
prometheus_client==0.21.0
propcache==0.2.0
psycopg==3.2.3
psycopg-c==3.2.3
pycparser==2.22
pydantic==2.8.1
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
redmail==0.6.0
rich==13.9.4
s3transfer==0.10.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tomli==2.2.1
typer==0.13.0
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
yarl==1.17.1
